+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - Add convenience classes for creating screens for the new navigation framework
 - [In Progress: 21 Dec 2020] Instant search in Faster registration
 - [In Progress: 23 Dec 2020] Migrate app to use ViewBinding
-- [In Progress: 29 Dec 2020] Add Instant search screen
+- [In Progress: 31 Dec 2020] Add Instant search screen
 
 ### Fixes
 - Fix BPpassport prefill value to have display value